Refactor tv android app into platfor-app(in preperation to add content-app) and refactor package names and other strings.
