+++ examples/tv-app/android/App/.idea/.name
@@ -1,1 +1,1 @@
-CHIPTVServer
+MatterTVAndroid
