+++ examples/tv-app/android/App/.idea/gradle.xml
@@ -10,7 +10,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/platform-app" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
