--- examples/tv-app/android/App/app/src/main/java/com/tcl/chip/chiptvserver/ChipTvServerApplication.java
@@ -1,23 +1,0 @@
-package com.tcl.chip.chiptvserver;
-
-import android.app.Application;
-import android.content.Intent;
-import android.os.Build;
-import com.tcl.chip.chiptvserver.service.MatterServantService;
-
-public class ChipTvServerApplication extends Application {
-  @Override
-  public void onCreate() {
-    super.onCreate();
-    startMatterServantService();
-  }
-
-  private void startMatterServantService() {
-    Intent intent = new Intent(this, MatterServantService.class);
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-      getApplicationContext().startForegroundService(intent);
-    } else {
-      startService(intent);
-    }
-  }
-}
