--- examples/tv-app/android/App/app/src/test/java/com/tcl/chip/chiptvslave/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package com.tcl.chip.chiptvserver;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-  @Test
-  public void addition_isCorrect() {
-    assertEquals(4, 2 + 2);
-  }
-}
