+++ examples/tv-app/android/App/platform-app/.gitignore
