+++ examples/tv-app/android/App/platform-app/libs/README.md
