+++ examples/tv-app/android/App/platform-app/libs/jniLibs/README.md
