+++ examples/tv-app/android/App/platform-app/proguard-rules.pro
