+++ examples/tv-app/android/App/platform-app/src/main/res/drawable/ic_launcher_background.xml
