+++ examples/tv-app/android/App/platform-app/src/main/res/mipmap-hdpi/ic_launcher.webp
