+++ examples/tv-app/android/App/platform-app/src/main/res/mipmap-mdpi/ic_launcher_round.webp
