+++ examples/tv-app/android/App/platform-app/src/main/res/mipmap-xhdpi/ic_launcher_round.webp
