+++ examples/tv-app/android/App/platform-app/src/main/res/mipmap-xxhdpi/ic_launcher.webp
