+++ examples/tv-app/android/App/platform-app/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
