+++ examples/tv-app/android/App/platform-app/src/main/res/values/colors.xml
