+++ examples/tv-app/android/App/platform-app/src/test/java/com/matter/tv/server/ExampleUnitTest.java
@@ -1,0 +1,17 @@
+package com.matter.tv.server;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+public class ExampleUnitTest {
+  @Test
+  public void addition_isCorrect() {
+    assertEquals(4, 2 + 2);
+  }
+}
