+++ examples/tv-app/android/App/settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = "CHIPTVServer"
-include ':app'
+rootProject.name = "MatterTVAndroid"
+include ':platform-app'
